# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meeting Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { MeetingStatus } from './meetingStatus';
import type { User } from './user';

export interface Meeting {
  id: number;
  title: string;
  date: string;
  time: string;
  status: MeetingStatus;
  /** @nullable */
  project?: string | null;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  location?: string | null;
  chairperson?: User;
  attendeeCount?: number;
  taskCount?: number;
  hasMinutes?: boolean;
  minutesApproved?: boolean;
  createdAt: string;
}