# lib/api-zod/src/generated/types/meetingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meeting Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingUpdate {
  title?: string;
  date?: string;
  time?: string;
  status?: string;
  project?: string;
  team?: string;
  location?: string;
  objectives?: string;
  /** @nullable */
  chairpersonId?: number | null;
  attendeeIds?: number[];
  agendaItems?: string[];
}