# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meeting Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  username: string;
  fullName: string;
  email: string;
  role: UserRole;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  avatar?: string | null;
  createdAt?: string;
}